EbolaPigAnalytics
Ebola is a rare and deadly disease caused by infection with a strain of
Ebola virus. The 2014 Ebola epidemic is the largest in history,
affecting multiple countries in West Africa. The metadata used in this
analysis spans a series of African countries affected by the deadly
Ebola virus. The analysis also includes the locality of the countries,
the category and the value.